+++ Release-Notes.md
@@ -1,5 +1,11 @@
 # Fastly_Cdn Release Notes
 
+## 1.2.34
+
+- Fix for serialization issue regarding old config data for GeoIP Country Mapping (Magento version above 2.2)
+- Added shell function for converting Fastly config data to JSON manually (Magento version above 2.2), executed by: fastly:format:serializetojson
+- Added shell function for converting Fastly config data to serialize format manually (Should be used only to revert changes made from fastly:format:serializetojson), executed by: fastly:format:jsontoserialize
+
 ## 1.2.33
 
 - Don't cache /customer/section/load. This works around core bug where Cache-Control headers are set to cache https://github.com/fastly/fastly-magento2/pull/111